My Text Adventure Game
Henry Walton Jones, Jr., better known as Indiana Jones, is a fictional archaeologist 
and protagonist of the franchise of the same name, created by George Lucas in homage 
to the action heroes of the film series of the 1930s. The character appears in four 
films directed by Steven Spielberg and a television series in novels, comic books, 
video games, and other media.

My game is not based on one of the Indiana Jones movies, but there are 
references to the movie: "Indiana Jones and the Last Crusade".